moving all mpibind logic into allocation
